test: consolidate MeterWidget tests and remove duplicate logic-based test
